+++ src/models/taskModel.ts
@@ -2,11 +2,45 @@ import { LogFile } from "./logFileModel";
 import { Gres } from "./gresModel";
 
 
+function parseFilename(
+    pattern: string,
+    jobid: number | string,
+    name: string,
+    user: string,
+    node: string,
+    stepid?: number,
+): string {
+    if (pattern.includes('\\')) { return pattern.replaceAll('\\', '') }
+    const replacements: Record<string, string> = {
+        '%A': `${jobid}`,  // Job array master job allocation number
+        '%a': '%a',  // Job array index number, not implemented for now
+        '%b': '%b',  // Job array index modulo 10, not implemented for now
+        '%J': stepid ? `${jobid}.${stepid}` : `${jobid}`,  // jobid.stepid
+        '%j': `${jobid}`,  // jobid
+        '%N': `${node}`,  // short hostname
+        '%n': '0',  // node identifier, not implemented for now
+        '%s': stepid?.toString() ?? '%s',  // stepid
+        '%t': '0',  // task identifier, not implemented for now
+        '%u': user,  // user name
+        '%x': name,  // job name
+    };
+
+    pattern = pattern.replace('%%', '\0');
+    pattern = pattern.replace(/%([A-Za-z])/g, (match, p1) => replacements[`%${p1}`] ?? match);
+    return pattern.replace('\0', '%');
+}
+
+
+
 export enum TaskState {
     R = "RUNNING",
     PD = "PENDING",
     CG = "COMPLETING",
 }
+
+export type TaskProperties = {
+    [K in keyof Task]: Task[K] extends LogFile ? string : Task[K];
+};
 
 export class Task {
     readonly jobid: number;
@@ -17,14 +51,14 @@ export class Task {
     readonly gres: Gres | null;
     readonly limit_time: string;
     runing_time: string;
-    readonly command: string;
+    readonly command: LogFile;
     readonly out_path: LogFile;
     readonly err_path: LogFile;
     reason: string;
     finished: boolean;
     //JobID,Name:255,Username:20,State:20,NodeList,Gres:50,TimeLimit,TimeUsed,Command:255,STDOUT:255,STDERR:255,Reason:100
     constructor(jobid: string, name: string, user: string, state: string, node: string, gres: string, limit_time: string, runing_time: string, command: string, out_path: string, err_path: string, reason: string)
-    constructor(jobid: number, name: string, user: string, state: string, node: string, gres: Gres | null, limit_time: string, runing_time: string, command: string, out_path: LogFile, err_path: LogFile, reason: string, finished: boolean)
+    constructor(jobid: number, name: string, user: string, state: string, node: string, gres: Gres | null, limit_time: string, runing_time: string, command: LogFile, out_path: LogFile, err_path: LogFile, reason: string, finished: boolean)
 
     constructor(
         jobid: number | string,
@@ -35,28 +69,28 @@ export class Task {
         gres: string | Gres | null,
         limit_time: string,
         runing_time: string,
-        command: string,
+        command: string | LogFile,
         out_path: string | LogFile,
         err_path: string | LogFile,
         reason: string,
         finished?: boolean,
     ) {
         this.jobid = typeof jobid === "string" ? parseInt(jobid) : jobid;
-        this.name = name.replace(/%j/i, jobid.toString());
+        this.name = name;
         this.user = user;
         this.state = state;
         this.node = node;
         this.gres = typeof gres === "string" ? (gres === 'N/A' ? null : new Gres(gres)) : gres;
         this.limit_time = limit_time;
         this.runing_time = runing_time;
-        this.command = command;
-        this.out_path = typeof out_path === "string" ? new LogFile(out_path.replace(/%j/i, jobid.toString())) : out_path;
-        this.err_path = typeof err_path === "string" ? new LogFile(err_path.replace(/%j/i, jobid.toString())) : err_path;
+        this.command = typeof command === "string" ? new LogFile(command) : command;
+        this.out_path = typeof out_path === "string" ? new LogFile(parseFilename(out_path, jobid, name, user, node)) : out_path;
+        this.err_path = typeof err_path === "string" ? new LogFile(parseFilename(err_path, jobid, name, user, node)) : err_path;
         this.reason = reason === 'None' ? '' : reason;
         this.finished = finished ?? false;
     }
 
-    public static fromObject(obj: Task): Task {
+    public static fromObject(obj: TaskProperties): Task {
         return new Task(
             obj.jobid,
             obj.name,
@@ -66,9 +100,9 @@ export class Task {
             obj.gres,
             obj.limit_time,
             obj.runing_time,
-            obj.command,
-            obj.out_path,
-            obj.err_path,
+            new LogFile(obj.command),
+            new LogFile(obj.out_path),
+            new LogFile(obj.err_path),
             obj.reason,
             obj.finished
         );